Mount workspace PVC in Che server
Mount the PVC used for workspaces to the Che server
to allow the server to create directory that is mounted
in the workspace. This is necessary because directories
created by kubernetes when mounting a subpath are created
with root permissions and cannot be modified by a non-root
user.

See: https://github.com/kubernetes/kubernetes/issues/41638

Signed-off-by: Angel Misevski <amisevsk@redhat.com>